% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_common_params}
\alias{mdl_common_params}
\title{mdl Common Parameters}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}

\item{document_id}{One or more document IDs. Get the document IDs with \code{\link[=mdl_documents]{mdl_documents()}}}

\item{group_name}{A group name. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}

\item{group_id}{A group ID. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}

\item{folder_name}{A folder name. Get folder names with \code{\link[=mdl_folders]{mdl_folders()}}.}

\item{folder_id}{A folder ID. Get folder names with \code{\link[=mdl_folders]{mdl_folders()}}.}

\item{max_objects}{The maximum number of objects (documents, folders, files...) to retreive}

\item{condition}{Either \code{NULL} (default) or a function taking an object (document,
folder, file, ...) represented as a list and returning a logical value.
The function will return the objects that have returned \code{TRUE}.}
}
\description{
These are parameters common to several mendeleyr functions
}
