% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_delete_document}
\alias{mdl_delete_document}
\title{Delete one or more documents}
\usage{
mdl_delete_document(token, document_id, show_progress = interactive())
}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}

\item{document_id}{One or more document IDs. Get the document IDs with \code{\link[=mdl_documents]{mdl_documents()}}}

\item{show_progress}{logical value to show a progress bar}
}
\description{
Delete one or more documents
}
