% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_delete_document_one}
\alias{mdl_delete_document_one}
\title{Delete a Mendeley document}
\usage{
mdl_delete_document_one(token, document_id)
}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}

\item{document_id}{One or more document IDs. Get the document IDs with \code{\link[=mdl_documents]{mdl_documents()}}}
}
\description{
Delete a Mendeley document
}
