% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_download_file}
\alias{mdl_download_file}
\title{Download a file}
\usage{
mdl_download_file(token, file_row, destfile = NULL)
}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}

\item{file_row}{A row from \code{\link[=mdl_files]{mdl_files()}} with the file we want to download}

\item{destfile}{The path where we want to save destfile, by default keeps
the original file name in the current working directory}
}
\description{
Download a file
}
