% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_files}
\alias{mdl_files}
\title{Retrieve the list of files attached to documents or groups}
\usage{
mdl_files(token, document_id = NULL, group_name = NULL, group_id = NULL)
}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}

\item{document_id}{One or more document IDs. Get the document IDs with \code{\link[=mdl_documents]{mdl_documents()}}}

\item{group_name}{A group name. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}

\item{group_id}{A group ID. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}
}
\description{
Retrieve the list of files attached to documents or groups
}
