% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_folder_info}
\alias{mdl_folder_info}
\title{Gets information from a Mendeley Folder}
\usage{
mdl_folder_info(token, folder_id = NULL, folder_name = NULL,
  group_id = NULL, group_name = NULL)
}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}

\item{folder_id}{A folder ID. Get folder names with \code{\link[=mdl_folders]{mdl_folders()}}.}

\item{folder_name}{A folder name. Get folder names with \code{\link[=mdl_folders]{mdl_folders()}}.}

\item{group_id}{A group ID. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}

\item{group_name}{A group name. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}
}
\description{
Gets information from a Mendeley Folder
}
