% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_groups}
\alias{mdl_groups}
\title{Gets all the group information}
\usage{
mdl_groups(token)
}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}
}
\description{
Gets all the group information
}
