% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_to_bibtex}
\alias{mdl_to_bibtex}
\title{Creates a bib file from a Mendeley folder or group}
\usage{
mdl_to_bibtex(token, folder_name = NULL, folder_id = NULL,
  group_name = NULL, group_id = NULL, bibfile = NULL)
}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}

\item{folder_name}{A folder name. Get folder names with \code{\link[=mdl_folders]{mdl_folders()}}.}

\item{folder_id}{A folder ID. Get folder names with \code{\link[=mdl_folders]{mdl_folders()}}.}

\item{group_name}{A group name. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}

\item{group_id}{A group ID. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}

\item{bibfile}{"*.bib" file to write the documents from folder_name}
}
\description{
Creates a bib file from a Mendeley folder or group
}
