% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_to_POSIXct}
\alias{mdl_to_POSIXct}
\title{Convert a date time text given by Mendeley into a POSIXct time}
\usage{
mdl_to_POSIXct(date_text)
}
\arguments{
\item{date_text}{a character vector with the date times as given by mendeley}
}
\value{
A POSIXct object with the date times converted
}
\description{
Convert a date time text given by Mendeley into a POSIXct time
}
