% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mendeley_conf}
\alias{mendeley_conf}
\alias{mdl_conf_new}
\alias{mdl_conf_save}
\alias{mdl_conf_load}
\title{Obtain and use client_id and client_secret}
\usage{
mdl_conf_new()

mdl_conf_save(client_id, client_secret, where = ".mendeley_conf.json")

mdl_conf_load(where = ".mendeley_conf.json")
}
\arguments{
\item{client_id}{The client id given by Mendeley when you create an app at http://dev.mendeley.com/myapps.html}

\item{client_secret}{The client secret given by Mendeley when you create an app}

\item{where}{File name where mendeleyr will store the client_id and client_secret}
}
\description{
Obtain and use client_id and client_secret
}
