% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_documents}
\alias{mdl_documents}
\title{Get the Document IDs from a given folder_id}
\usage{
mdl_documents(token, folder_name = NULL, folder_id = NULL,
  group_name = NULL, group_id = NULL, modified_since = NULL,
  max_objects = Inf, condition = NULL)
}
\arguments{
\item{token}{A Token given by \code{\link[=mdl_token]{mdl_token()}} that provides access to your account.}

\item{folder_name}{A folder name. Get folder names with \code{\link[=mdl_folders]{mdl_folders()}}.}

\item{folder_id}{A folder ID. Get folder names with \code{\link[=mdl_folders]{mdl_folders()}}.}

\item{group_name}{A group name. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}

\item{group_id}{A group ID. Read group IDs through \code{\link[=mdl_groups]{mdl_groups()}}}

\item{modified_since}{Only retrieve documents modified since that date.}

\item{max_objects}{The maximum number of objects (documents, folders, files...) to retreive}

\item{condition}{Either \code{NULL} (default) or a function taking an object (document,
folder, file, ...) represented as a list and returning a logical value.
The function will return the objects that have returned \code{TRUE}.}
}
\description{
Get the Document IDs from a given folder_id
}
