% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeleyr.R
\name{mdl_token}
\alias{mdl_token}
\title{Obtains a Mendeley Token valid for the current session.}
\usage{
mdl_token(mendeley_conf)
}
\arguments{
\item{mendeley_conf}{a list with two items: \code{client_id} and \code{client_secret}.}
}
\description{
The first time you get the token you will need to authorise your application
(mendeleyr) access to your account.
}
\examples{
\dontrun{
token <- mdl_token(mdl_conf_load())
}
}
